Better Pickups and UI
Added a Canvas for showing the number of lives of the player and also his points. Pickups don't respawn everytime the player dies. TODO - better UI for points and lives. SFX and one main song for each level, a Level 2 and maybe a Level 3, give the player the possibility to attack and destroy enemies, enemies that attacks the player.